Rename obsolete interfaces
