+++ ProcessDoctor.Tests/ProcessTreeViewModelTests.cs
@@ -1,7 +1,6 @@
 using System.Collections.ObjectModel;
 using JetBrains.Lifetimes;
 using ProcessDoctor.Backend.Core;
-using ProcessDoctor.TestFramework;
 using ProcessDoctor.TestFramework.Logging;
 using ProcessDoctor.ViewModels;
 using Xunit.Abstractions;
